Initial implementation of SMIOL_open_file and SMIOL_close_file
The SMIOL_open_file routine now allocates a file handle and saves a pointer
to the SMIOL_context in that handle, and the SMIOL_close_file routine
deallocates all memory associated with a SMIOL_file. The Fortran routines
pass valid arguments to the C open/close routines, and check the return
values from the C routines.

No files are actually created or opened by the SMIOL_open_file routine, yet.
It is expected that, once support for a library like parallel-netCDF is added,
the implementations of SMIOL_open_file and SMIOL_close_file will actually
affect the state of the filesystem.