Provide initial C unit tests for SMIOL_put_var
The unit tests implemented in this commit verify that the return codes from
SMIOL_put_var match the expected values. But, without the ability to read data
back from a file, it is not yet possible to verify that values were written
correctly. Manual inspection of the contents of the new test file,
test_put_get_vars.nc, indicates that data are being written correctly at
present.