Run the async_write function from its own thread
The async_write function was previously called directly from the SMIOL_put_var
function. Now, async_write is run in its own thread, which is started by a call
to SMIOL_async_launch_thread. However, the call to SMIOL_async_launch_thread is
immediately followed by a call to SMIOL_async_join_thread, so writing of data
to a file is still a blocking operation.