Avoid trying to free an invalid aggregation communicator in SMIOL_free_decomp
If the SMIOL_decomp struct that is passed to the SMIOL_free_decomp function was
not produced by a call to SMIOL_create_decomp (which can happen, for example, if
the decomp was produced by a direct call to build_exchange), the aggregation
members of the decomp may be uninitialized, and this may result in an attempt to
free an invalid communicator with MPI_Comm_free.

To avoid problems in SMIOL_free_decomp, all aggregation-specific members of
the SMIOL_decomp struct are initialized in build_exchange. In particular,
the agg_comm member is initialized to MPI_COMM_NULL. In the SMIOL_free_decomp
function, the agg_comm is only freed if it is not MPI_COMM_NULL.