num_io_tasks and io_stride are now part of SMIOL_context
Prior to this commit, values for num_io_tasks and io_stride were provided in
the call to SMIOL_create_decomp. This provided the flexibility to have different
sets of tasks reading from or writing to the same file when variables used
decompositions with different num_io_tasks and io_stride values.

In practice, however, all decompositions created by MPAS will use the same
num_io_tasks and io_stride values. Furthermore, being able to guarantee that all
variables in a file are read or written with the same set of I/O tasks may
provide opportunities for optimization.

With this commit, the number of I/O tasks and the stride between those tasks
are specified when a SMIOL context is created, and all decompositions created
within the same context inherit those values.